add event buffer criterion benchmarks
