+++ bd-event-buffer/Cargo.toml
@@ -16,6 +16,7 @@ time.workspace              = true
 tokio.workspace             = true
 
 [dev-dependencies]
+criterion          = "0.8.2"
 proptest.workspace = true
 
 [target.'cfg(target_os = "linux")'.dev-dependencies]
@@ -24,3 +25,7 @@ gungraun = { version = "0.19.4", features = ["client_requests"] }
 [[bench]]
 harness = false
 name    = "main"
+
+[[bench]]
+harness = false
+name    = "criterion"
